+++ src/kel/__init__.py
@@ -44,4 +44,4 @@ __all__ = [
     "get_tracer",
 ]
 
-__version__ = "1.4.0"
+__version__ = "1.5.0"